backend/requirements.txt
flask
flask-cors
pandas
scikit-learn
numpy
google-generativeai
python-dotenv
